+++ std/pgetopt.C
@@ -121,7 +121,7 @@ local inline char *get_version(const char *cvs_id)
     while (*end > ' ') end++;
 
     if (*end == ' ') {
-	std::ptrdiff_t n = end-start+1;
+	std::size_t n = static_cast<std::size_t>(end-start+1);
 	char *version = new char[n];
 	strncpy(version, start, n-1);
 	version[n-1] = 0;
@@ -153,7 +153,7 @@ local inline char *get_name(const char *source)
 
     if (end < start) return NULL;
 
-    std::ptrdiff_t n = end-start+2;
+    std::size_t n = static_cast<std::size_t>(end-start+2);
     char *name = new char[n];
     strncpy(name, start, n-1);
     name[n-1] = 0;
